add human-readable titles for prefixed quicksearches
